requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3
Django==2.0.7
django-js-asset==1.1.0
django-markdown==0.8.4
django-mdeditor==0.1.6
idna==2.7
Markdown==2.6.11
markdown2==2.3.5
pycparser==2.18
Pygments==2.2.0
PyMySQL==0.9.2
pytz==2018.5
six==1.11.0
